requirements.txt
anthropic>=0.25.0
yfinance>=0.2.38
ccxt>=4.3.0
alpaca-py>=0.20.0
pandas>=2.2.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.1
ta>=0.11.0
colorama>=0.4.6
